% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{RIVER}{priver fish was found in}
  \item{MILE}{mile fish was found at}
  \item{SPECIES}{species of fish}
  \item{LENGTH}{length of fish}
  \item{WEIGHT}{weight of fish}
  \item{DDT}{ddt of fish}
}
}
\source{
Textbook
}
\usage{
ddt
}
\description{
144 lines of measurements for assorted fish
}
\keyword{datasets}
