% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2().R
\name{myboot2}
\alias{myboot2}
\title{My Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{population vector}

\item{fun}{function to be used}

\item{alpha}{confidence interval}

\item{cx}{}
}
\value{
a plot of the mean setimates and interval estimates on the sample
}
\description{
My Bootstrap Function
}
\examples{
\dontrun {
myboot2(iter = 10000, x = sam, fun = "mean", alpha = 0.05)
}
}
