% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt function}
\usage{
myddt(df, species)
}
\arguments{
\item{df}{datqframe}

\item{spec}{chosen species}
}
\value{
a plot of the chosen species, the sf before and after subsetting,
and a relative frequency table of rivers
}
\description{
myddt function
}
