% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
a histogram of the sum of uniforms
}
\description{
mycltu
}
\examples{
dont run{
mycltu(n = 5, iter = 10000, a = 0, b = 10)
}
}
