% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt_2.R
\name{myddt}
\alias{myddt}
\title{myddt() for project 1}
\usage{
myddt(df, x, y, SPECIES, col)
}
\arguments{
\item{df}{ddt datafrane}

\item{x}{length}

\item{y}{weight}

\item{SPECIES}{chosen species}

\item{col}{rivers}
}
\value{
a scatter plot, two csv files, and a relative frequency table
}
\description{
myddt() for project 1
}
\examples{
\dontrun{
myddt(ddt, "LENGTH", "WEIGHT", "CCATFISH", "RIVER")}
}
